fix - applied & suggested project card
지원한, 제안받은 카드 디자인 통일
프로젝트카드컨텐츠 컴포넌트로 분리
